+++ projects/miopen/src/include/miopen/softmax/problem_description.hpp
@@ -106,16 +106,16 @@ struct MIOPEN_INTERNALS_EXPORT ProblemDescription : ProblemDescriptionBase,
     NetworkConfig MakeNetworkConfig() const override;
 
     template <class Self>
-    static void Visit(Self&& self, std::function<void(int64_t, std::string)> f)
+    static void Visit(Self&& self, std::function<void(uint64_t, std::string)> f)
     {
         // The column names match the driver command line argument names
-        f(static_cast<uint64_t>(self.isForward), "forw");
+        f(self.isForward, "forw");
         f(self.GetBatchSize(), "batchsize");
         f(self.GetChannels(), "in_channels");
         f(self.GetHeight(), "in_h");
         f(self.GetWidth(), "in_w");
-        f(static_cast<uint64_t>(self.algorithm), "algorithm");
-        f(static_cast<uint64_t>(self.mode), "mode");
+        f(self.algorithm, "algorithm");
+        f(self.mode, "mode");
     }
 
     template <class Self>
@@ -128,7 +128,7 @@ struct MIOPEN_INTERNALS_EXPORT ProblemDescription : ProblemDescriptionBase,
     template <class Self, class Visitor>
     static void VisitAll(Self&& self, const Visitor& f)
     {
-        Visit(std::forward<Self>(self), [&](int64_t value, std::string name) { f(value, name); });
+        Visit(std::forward<Self>(self), [&](uint64_t value, std::string name) { f(value, name); });
         Visit(std::forward<Self>(self),
               [&](std::string value, std::string name) { f(value, name); });
     }
